pyproject.toml
[project]
name = "langchain-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.7",
    "langchain-classic>=1.0.1",
    "langchain-google-genai>=4.2.0",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
]
